adding readme, caching and allebestellingen page

+++ static/script/cacheStap2.js
@@ -1,5 +1,9 @@
-var orderButton = document.querySelector("#order")
+if (typeof Storage !== void 0) {
+  var orderButton = document.querySelector("#order")
 
-orderButton.addEventListener("click", function () {
-  localStorage.clear()
-})
+  orderButton.addEventListener("click", function () {
+    localStorage.clear()
+  })
+} else {
+  console.log("No localStorage supported")
+}
